requirements.txt
pytest
PyMySQL
allure-pytest
requests
PyYAML
pytest-html
redis